+++ pkg/plugins/plugins_integration_test.go
@@ -12,6 +12,8 @@ import (
 )
 
 func TestIntegrationPlugins(t *testing.T) {
+	t.Skip("Skipping `Plugins` integration test due API reaching EOL (returns 410 HTTP status)")
+
 	t.Parallel()
 
 	tc := mock.NewIntegrationTestConfig(t)