+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+<a name="v0.75.0"></a>
+## [v0.75.0] - 2022-04-13
+### Features
+- **errors:** handle 402 payment required HTTP response scenario
+
+<a name="v0.74.2"></a>
+## [v0.74.2] - 2022-03-23
+### Bug Fixes
+- use correct input type for cloud disable integrations mutation
+
+<a name="v0.74.1"></a>
+## [v0.74.1] - 2022-03-04
+### Bug Fixes
+- remove integrations from getLinkedAccounts query
+
+<a name="v0.74.0"></a>
+## [v0.74.0] - 2022-03-03
+### Features
+- **auth:** Add X-Account-ID header if value exists in request context
+
 <a name="v0.73.0"></a>
 ## [v0.73.0] - 2022-02-09
 ### Features
@@ -1246,7 +1266,11 @@
 - extract paging implementation
 - rename packages for clarity, promote Config to the public package
 
-[Unreleased]: https://github.com/newrelic/newrelic-client-go/compare/v0.73.0...HEAD
+[Unreleased]: https://github.com/newrelic/newrelic-client-go/compare/v0.75.0...HEAD
+[v0.75.0]: https://github.com/newrelic/newrelic-client-go/compare/v0.74.2...v0.75.0
+[v0.74.2]: https://github.com/newrelic/newrelic-client-go/compare/v0.74.1...v0.74.2
+[v0.74.1]: https://github.com/newrelic/newrelic-client-go/compare/v0.74.0...v0.74.1
+[v0.74.0]: https://github.com/newrelic/newrelic-client-go/compare/v0.73.0...v0.74.0
 [v0.73.0]: https://github.com/newrelic/newrelic-client-go/compare/v0.72.0...v0.73.0
 [v0.72.0]: https://github.com/newrelic/newrelic-client-go/compare/v0.71.0...v0.72.0
 [v0.71.0]: https://github.com/newrelic/newrelic-client-go/compare/v0.70.0...v0.71.0